requirements.txt
Django==1.10
MySQL-python==1.2.5
Pillow==3.4.1
argparse==1.2.1
django-crispy-forms==1.6.0
django-simple-captcha==0.5.3
djangorestframework==3.4.4
six==1.10.0
unittest-xml-reporting==2.1.0
wsgiref==0.1.2
gunicorn==19.3.0